Refactor losses
