+++ code/models/endd.py
@@ -1,4 +1,4 @@
-from utils.DirichletEnDDLoss import DirichletEnDDLoss
+from utils import losses
 
 def get_model(base_model):
     """Take an uncompiled model and return model compiled for ENDD.
@@ -6,5 +6,5 @@ def get_model(base_model):
     Warning: This function works in place. Model is returned only for
     conveniance.
     """
-    base_model.compile(optimizer='adam', loss=DirichletEnDDLoss())
+    base_model.compile(optimizer='adam', loss=losses.DirichletEnDDLoss())
     return base_model
