+++ code/scripts/priornet_cifar.py
@@ -11,11 +11,8 @@ import tensorflow as tf
 from tensorflow.keras import datasets
 
 import settings
-from utils.simplex import plot_simplex
-from utils.DirichletKL import DirichletKL
-from models.cnn_priorNet import get_model
-from utils import preprocessing
-from utils import saveload
+from utils import preprocessing, saveload, simplex, losses
+from models import cnn_priorNet
 
 DATASET = 'cifar10_PN'
 PLOT_SIMPLEX = False
@@ -26,7 +23,7 @@ EPOCHS = 2
 (train_images, train_labels), (test_images, test_labels) = datasets.cifar10.load_data()
 train_images, train_logits, test_images, test_logits = preprocessing.preprocess_cifar_for_priornet(
     train_images, train_labels, test_images, test_labels)
-model = get_model(DATASET)
+model = cnn_priorNet.get_model(DATASET)
 model.fit(train_images, train_logits, batch_size=BATCH_SIZE, epochs=EPOCHS)
 if SAVE_WEIGHTS:
     saveload.save_tf_model(model, "cnn_priorNet")
@@ -37,4 +34,4 @@ real = tf.math.argmax(tf.squeeze(test_logits), axis=1)
 score = tf.math.reduce_sum(tf.cast(predictions == real, tf.float32)) / len(real)
 print('score: ' + str(score))
 if PLOT_SIMPLEX:
-    plot_simplex(logits)
+    simplex.plot_simplex(logits)
