+++ code/utils/losses.py
@@ -1,7 +1,8 @@
-
 import tensorflow as tf
-from tensorflow.math import exp,reduce_sum,lgamma,reduce_mean,log
+from tensorflow.math import exp, reduce_sum, lgamma, reduce_mean, log, digamma
 from tensorflow.nn import softmax
+
+
 class DirichletEnDDLoss(tf.keras.losses.Loss):
     """
     Negative Log-likelihood of the model on the transfer dataset"""
@@ -35,3 +36,28 @@ class DirichletEnDDLoss(tf.keras.losses.Loss):
         cost = target_dependent_term + target_independent_term
 
         return reduce_mean(cost) * (temp ** 2) #mean of all batches
+
+
+class DirichletKL(tf.keras.losses.Loss):
+    '''
+      This function computes the Reverse KL divergence between two dirichlet distributions based on their alphas.
+        Inputs:
+              alpha_true (Batch_size x n_classes), the alphas corresponding to the true distribution. (1,1,1) for OOD and (100,1,1) / (1,100,1) / (1,1,100) for ID.
+              alpha_pred (Batch_size x n_classes), the predicted alphas from our network given input X.
+              epsilon = 1e-8, smoothing factor
+        Output:
+              KL-divergences (Batch-Size x 1), Reverse KL divergence of the two dirichlet distributions RKL(true_dirichlet||pred_dirichlet)
+    '''
+    def __init__(self, epsilon=1e-10):
+        super().__init__()
+        self.epsilon = epsilon
+
+    def call(self, alpha_true, alpha_pred):
+        epsilon = self.epsilon
+        KL = lgamma(tf.math.reduce_sum(alpha_pred)) - tf.math.reduce_sum(
+            lgamma(alpha_pred + epsilon)) - lgamma(
+                tf.math.reduce_sum(alpha_true)) + tf.math.reduce_sum(
+                    lgamma(alpha_true + epsilon)) + tf.math.reduce_sum(
+                        (alpha_pred - alpha_true) *
+                        (digamma(alpha_pred + epsilon) - digamma(tf.math.reduce_sum(alpha_pred))))
+        return KL
